+++ src/ControlDBServiceProvider.php
@@ -21,6 +21,14 @@ use BristolSU\ControlDB\Contracts\Repositories\Pivots\Tags\RoleRoleTag as RoleRo
 use BristolSU\ControlDB\Contracts\Repositories\Pivots\Tags\PositionPositionTag as PositionPositionTagContract;
 use BristolSU\ControlDB\Contracts\Repositories\Pivots\UserGroup as UserGroupContract;
 use BristolSU\ControlDB\Contracts\Repositories\Pivots\UserRole as UserRoleContract;
+use BristolSU\ControlDB\Events\DataGroup\DataGroupEventDispatcher;
+use BristolSU\ControlDB\Events\DataPosition\DataPositionEventDispatcher;
+use BristolSU\ControlDB\Events\DataRole\DataRoleEventDispatcher;
+use BristolSU\ControlDB\Events\DataUser\DataUserEventDispatcher;
+use BristolSU\ControlDB\Events\User\UserEventDispatcher;
+use BristolSU\ControlDB\Events\Group\GroupEventDispatcher;
+use BristolSU\ControlDB\Events\Role\RoleEventDispatcher;
+use BristolSU\ControlDB\Events\Position\PositionEventDispatcher;
 use BristolSU\ControlDB\Export\ExportControlCommand;
 use BristolSU\ControlDB\Export\ExportManager;
 use BristolSU\ControlDB\Models\DataUser;
@@ -121,8 +129,6 @@ use BristolSU\ControlDB\Contracts\Repositories\Tags\RoleTagCategory as RoleTagCa
 use BristolSU\ControlDB\Contracts\Repositories\Tags\UserTag as UserTagRepositoryContract;
 use BristolSU\ControlDB\Contracts\Repositories\Tags\UserTagCategory as UserTagCategoryRepositoryContract;
 use BristolSU\ControlDB\Contracts\Repositories\User as UserRepositoryContract;
-use Illuminate\Contracts\Container\BindingResolutionException;
-use Illuminate\Database\Eloquent\Factory;
 use Illuminate\Support\Facades\Route;
 use Illuminate\Support\ServiceProvider;
 
@@ -148,6 +154,19 @@ class ControlDBServiceProvider extends ServiceProvider
         $this->setupRouteModelBinding();
         $this->setupRoutes();
         $this->setupObservers();
+        $this->registerModelEvents();
+    }
+
+    public function registerModelEvents()
+    {
+        $this->app->extend(DataUserRepositoryContract::class, fn (DataUserRepositoryContract $service) => new DataUserEventDispatcher($service));
+        $this->app->extend(UserRepositoryContract::class, fn (UserRepositoryContract $service) => new UserEventDispatcher($service));
+        $this->app->extend(DataGroupRepositoryContract::class, fn (DataGroupRepositoryContract $service) => new DataGroupEventDispatcher($service));
+        $this->app->extend(GroupRepositoryContract::class, fn (GroupRepositoryContract $service) => new GroupEventDispatcher($service));
+        $this->app->extend(DataRoleRepositoryContract::class, fn (DataRoleRepositoryContract $service) => new DataRoleEventDispatcher($service));
+        $this->app->extend(RoleRepositoryContract::class, fn (RoleRepositoryContract $service) => new RoleEventDispatcher($service));
+        $this->app->extend(DataPositionRepositoryContract::class, fn (DataPositionRepositoryContract $service) => new DataPositionEventDispatcher($service));
+        $this->app->extend(PositionRepositoryContract::class, fn (PositionRepositoryContract $service) => new PositionEventDispatcher($service));
     }
 
     /**